oss_discovery_engine/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
